Changes in the previous version I forgot to add: Reversed EntityType to have Character be the first to make the sorting algorithm work properly Added a debug print to console Moved the listeners in gameMapImpl to a private method "Init" and the entity creation bit in "addEntities" to keep the constructor clean.
Changes in the current version.
Made the chat panel twice as wide, added a border to show what it's for, renamed directions and moveTo method in GameMapImpl.
Added the space key to the accepted actions in order to let you wait

TODO: add world singleton, make the GUI a little nicer, change the controllers to be static nested classes, add win and lose screens, add a small panel with all the available commands, also add smart enemies that follow the player around