Started changing the visibility of some classes Changed the main to be cleaner, since singletons don't require active initialization (check "lazy initialization", when they are asked for the first time, they get created) Changed the title of the game in the MainFrame Changed the chatLog to have clear text Changed Controller, GameController and GameChatController to use the singleton design pattern Changed GamePanel to have a cleaner constructor by adding an init method, this is subject to change as it's not working properly yet. (Look at TO-DOs)
Added a way to win and lose in GameController. This is accessible statically too so I no longer need to pass it through parameters and constructors.
NOTE: Some variables still hold the instance of the controller singletons, this was done to make the code shorter and be less verbose

TODO: add world singleton, make the GUI a little nicer, change the controllers to be static nested classes, finish win and lose screens, add a small panel with all the available commands, also add smart enemies that follow the player around

KNOWN ISSUES: Winning and losing still doesn't change your screen, must be finished.
Entities should first override the toString method for when they are required in the chatLog