Changed Character to increase its score when it removes an enemy Changed ChatArea to have fixed dimensions, changed its listener to let it receive messages and scores Changed GameArea to have fixed dimensions
Added Counter Interface, this is used to keep track of the player's score (CounterImpl added to define an use for said counter
Changed Pair to an interface and made PairImpl its implementation
Changed Enemy to send the score when you get eaten
Removed Enemy sending a message to the chat as it's not needed
Changed GameChatController to let it update the score
Changed finishGame method to show you your final score
Changed GameLogic to have a singular counter for the score NOTE: Maybe there's a better way
Changed MapReader to read any size of .txt as long as it's given in a matrix.
NOTE: There are no checks that make sure the map was well-defined, also must change the way everything is initialized now to actually make it work. USE A FINAL
Changed setToGameOver to use the score too
Changed the usable keys to be the direction arrows instead of WASD
Changed GamePanel to use the new scoring system in the final screen
NOTE: it shows in the top part which isn't ideal.
Changed Map_1

TODO: add world singleton, make the GUI a little nicer, change the controllers to be static nested classes, finish win and lose screens, add a small panel with all the available commands, also add smart enemies that follow the player around